Fix current_time in legacy timestamp semantics
Fix `current_time` in legacy timestamp semantics when session zone has
different zone offset on current date compared to offset on 1970-01-01.
